misc: hasitem matcher was moved from test file
